Implemented the basic functionality for the activate switch. Moved announcement classes to their own file. Renamed bootstrap folder
